Reflect review comments
